Add files via upload
Congratulations on getting to this point of the course🎉! The project is an exciting exercise where you get to put into practice all the concepts you have been taught so far. As discussed in class, please use the resources in the attached folder as a guide on how you will make your presentations. Additionally, kindly take note of the following:
.Your submission should be in the form of a zipped folder. It should contain your:
Notebook (.ipynb) and PowerPoint presentation (.pptx) OR 
Dashboard (.pbix) file and Notebook (.ipynb) OR 
Dashboard (.pbix) file and PowerPoint presentation (.pptx)
A presentation schedule will be availed to you, this will act as a guide to show who will present when and at what time.
Each presentation will take a maximum of 10 minutes (7 minutes presentation, 3 minutes Q&A). This should be enough time for the student to present their work and have a Q&A session with the audience as well.